The module code is now partially rebuilt, once it is done focus on Eolian automation will resume.

+++ src/lib/global/type/angl.h
@@ -8,7 +8,7 @@ typedef BASE Angl;
 
 #define AnglSOA Eina_Inarray*
 
-inline void AnglSIZE(uint32_t *counter)
+inline void AnglSIZE(uint64_t *counter)
 {
    *counter += 1;
 }
@@ -33,17 +33,17 @@ inline void AnglPOP(Eina_Inarray **key)
    eina_inarray_pop(*key);
 }
 
-inline void AnglREAD(Eina_Inarray **key, Angl *bufferkey, uint32_t index)
+inline void AnglREAD(Eina_Inarray **key, Angl *bufferkey, uint64_t index)
 {
    *bufferkey = *(Angl*)eina_inarray_nth(*key, index);
 }
 
-inline void AnglWRITE(Eina_Inarray **key, Angl *bufferkey, uint32_t index)
+inline void AnglWRITE(Eina_Inarray **key, Angl *bufferkey, uint64_t index)
 {
    eina_inarray_replace_at(*key, index, bufferkey);
 }
 
-inline void AnglCOPY(Eina_Inarray **source, Eina_Inarray **destination, uint32_t index)
+inline void AnglCOPY(Eina_Inarray **source, Eina_Inarray **destination, uint64_t index)
 {
    eina_inarray_replace_at(*source, index, *destination);
 }
